requirements.txt
google-generativeai
pandas
openpyxl
python-dotenv
langchain
transformers
torch
sentence-transformers
faiss-cpu
numpy